+++ crates/model/src/data/mod.rs
@@ -109,8 +109,8 @@ pub enum Data {
     MarkPriceUpdate(MarkPriceUpdate), // TODO: Rename to MarkPrice once Cython gone
     IndexPriceUpdate(IndexPriceUpdate), // TODO: Rename to IndexPrice once Cython gone
     FundingRateUpdate(FundingRateUpdate),
-    InstrumentStatus(InstrumentStatus),
     OptionGreeks(OptionGreeks),
+    InstrumentStatus(InstrumentStatus),
     InstrumentClose(InstrumentClose),
     Custom(CustomData),
     #[cfg(feature = "defi")]
@@ -154,11 +154,11 @@ impl TryFrom<Data> for DataFFI {
             Data::FundingRateUpdate(_) => {
                 anyhow::bail!("Cannot convert Data::FundingRateUpdate to DataFFI")
             }
-            Data::InstrumentStatus(_) => {
-                anyhow::bail!("Cannot convert Data::InstrumentStatus to DataFFI")
-            }
             Data::OptionGreeks(_) => {
                 anyhow::bail!("Cannot convert Data::OptionGreeks to DataFFI")
+            }
+            Data::InstrumentStatus(_) => {
+                anyhow::bail!("Cannot convert Data::InstrumentStatus to DataFFI")
             }
             Data::InstrumentClose(x) => Ok(Self::InstrumentClose(x)),
             Data::Custom(_) => anyhow::bail!("Cannot convert Data::Custom to DataFFI"),
@@ -226,10 +226,10 @@ impl<'de> Deserialize<'de> for Data {
             "FundingRateUpdate" => Ok(Self::FundingRateUpdate(
                 serde_json::from_value(value).map_err(D::Error::custom)?,
             )),
-            "InstrumentStatus" => Ok(Self::InstrumentStatus(
+            "OptionGreeks" => Ok(Self::OptionGreeks(
                 serde_json::from_value(value).map_err(D::Error::custom)?,
             )),
-            "OptionGreeks" => Ok(Self::OptionGreeks(
+            "InstrumentStatus" => Ok(Self::InstrumentStatus(
                 serde_json::from_value(value).map_err(D::Error::custom)?,
             )),
             "InstrumentClose" => Ok(Self::InstrumentClose(
@@ -260,8 +260,8 @@ impl Clone for Data {
             Self::MarkPriceUpdate(x) => Self::MarkPriceUpdate(*x),
             Self::IndexPriceUpdate(x) => Self::IndexPriceUpdate(*x),
             Self::FundingRateUpdate(x) => Self::FundingRateUpdate(*x),
-            Self::InstrumentStatus(x) => Self::InstrumentStatus(*x),
             Self::OptionGreeks(x) => Self::OptionGreeks(*x),
+            Self::InstrumentStatus(x) => Self::InstrumentStatus(*x),
             Self::InstrumentClose(x) => Self::InstrumentClose(*x),
             Self::Custom(x) => Self::Custom(x.clone()),
             #[cfg(feature = "defi")]
@@ -282,8 +282,8 @@ impl PartialEq for Data {
             (Self::MarkPriceUpdate(a), Self::MarkPriceUpdate(b)) => a == b,
             (Self::IndexPriceUpdate(a), Self::IndexPriceUpdate(b)) => a == b,
             (Self::FundingRateUpdate(a), Self::FundingRateUpdate(b)) => a == b,
-            (Self::InstrumentStatus(a), Self::InstrumentStatus(b)) => a == b,
             (Self::OptionGreeks(a), Self::OptionGreeks(b)) => a == b,
+            (Self::InstrumentStatus(a), Self::InstrumentStatus(b)) => a == b,
             (Self::InstrumentClose(a), Self::InstrumentClose(b)) => a == b,
             (Self::Custom(a), Self::Custom(b)) => a == b,
             #[cfg(feature = "defi")]
@@ -308,8 +308,8 @@ impl Serialize for Data {
             Self::MarkPriceUpdate(x) => x.serialize(serializer),
             Self::IndexPriceUpdate(x) => x.serialize(serializer),
             Self::FundingRateUpdate(x) => x.serialize(serializer),
-            Self::InstrumentStatus(x) => x.serialize(serializer),
             Self::OptionGreeks(x) => x.serialize(serializer),
+            Self::InstrumentStatus(x) => x.serialize(serializer),
             Self::InstrumentClose(x) => x.serialize(serializer),
             Self::Custom(x) => x.serialize(serializer),
             #[cfg(feature = "defi")]
@@ -354,8 +354,8 @@ impl_try_from_data!(Bar, Bar);
 impl_try_from_data!(MarkPriceUpdate, MarkPriceUpdate);
 impl_try_from_data!(IndexPriceUpdate, IndexPriceUpdate);
 impl_try_from_data!(FundingRateUpdate, FundingRateUpdate);
-impl_try_from_data!(InstrumentStatus, InstrumentStatus);
 impl_try_from_data!(OptionGreeks, OptionGreeks);
+impl_try_from_data!(InstrumentStatus, InstrumentStatus);
 impl_try_from_data!(InstrumentClose, InstrumentClose);
 
 /// Converts a vector of `Data` items to a specific variant type.
@@ -383,8 +383,8 @@ impl Data {
             Self::MarkPriceUpdate(mark_price) => mark_price.instrument_id,
             Self::IndexPriceUpdate(index_price) => index_price.instrument_id,
             Self::FundingRateUpdate(funding_rate) => funding_rate.instrument_id,
-            Self::InstrumentStatus(status) => status.instrument_id,
             Self::OptionGreeks(greeks) => greeks.instrument_id,
+            Self::InstrumentStatus(status) => status.instrument_id,
             Self::InstrumentClose(close) => close.instrument_id,
             Self::Custom(custom) => custom
                 .data_type
@@ -455,8 +455,8 @@ impl_catalog_path_prefix!(Bar, "bars");
 impl_catalog_path_prefix!(IndexPriceUpdate, "index_prices");
 impl_catalog_path_prefix!(MarkPriceUpdate, "mark_prices");
 impl_catalog_path_prefix!(FundingRateUpdate, "funding_rate_update");
-impl_catalog_path_prefix!(InstrumentStatus, "instrument_status");
 impl_catalog_path_prefix!(OptionGreeks, "option_greeks");
+impl_catalog_path_prefix!(InstrumentStatus, "instrument_status");
 impl_catalog_path_prefix!(InstrumentClose, "instrument_closes");
 
 use crate::instruments::InstrumentAny;
@@ -474,8 +474,8 @@ impl HasTsInit for Data {
             Self::MarkPriceUpdate(p) => p.ts_init,
             Self::IndexPriceUpdate(p) => p.ts_init,
             Self::FundingRateUpdate(f) => f.ts_init,
-            Self::InstrumentStatus(s) => s.ts_init,
             Self::OptionGreeks(g) => g.ts_init,
+            Self::InstrumentStatus(s) => s.ts_init,
             Self::InstrumentClose(c) => c.ts_init,
             Self::Custom(c) => c.data.ts_init(),
             #[cfg(feature = "defi")]
@@ -546,15 +546,15 @@ impl From<FundingRateUpdate> for Data {
     }
 }
 
-impl From<InstrumentStatus> for Data {
-    fn from(value: InstrumentStatus) -> Self {
-        Self::InstrumentStatus(value)
-    }
-}
-
 impl From<OptionGreeks> for Data {
     fn from(value: OptionGreeks) -> Self {
         Self::OptionGreeks(value)
+    }
+}
+
+impl From<InstrumentStatus> for Data {
+    fn from(value: InstrumentStatus) -> Self {
+        Self::InstrumentStatus(value)
     }
 }